Make regex matching case insensitive following the critical use cases.
It's unlikely FOO and foo has the different payee,
while case can be changed arbitrarily by connected system.
In future we can add case-sensitivity config if needed...

The currrent structure is useful so that we can configure EntityMatcher
for each format, but this flexibility exposes code duplication for this
change. It'd also arise when we introduce other based matcher.
Potentially we can stop using GAT-based polymorphism and get back to
more simpler approach (e.g. provide EnumMatcher, RegexMatcher and
translate it with simple args) so that we can use unified matcher logic.
This would help us to fuse regex or matchers into one gigantic regex,
which in term much optimized.